requirements.txt
python-dotenv==0.10.1
Flask==1.0.2
Flask-WTF==0.14.2
pymongo
bson
